added forward/previous page logic to GUI
