--- dndBuilder/src/GUI/Panels/DescriptionPanel.java
@@ -1,4 +1,0 @@
-package GUI.Panels;
-
-public class DescriptionPanel {
-}
