--- dndBuilder/src/GUI/Panels/EquipmentPanel.java
@@ -1,4 +1,0 @@
-package GUI.Panels;
-
-public class EquipmentPanel {
-}
