--- dndBuilder/src/GUI/Panels/FinalizePanel.java
@@ -1,4 +1,0 @@
-package GUI.Panels;
-
-public class FinalizePanel {
-}
